payment_service/pyproject.toml
[tool.poetry]
name = "payment-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.8"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
drf-spectacular = "^0.27.2"
stripe = "^10.5.0"
django-cors-headers = "^4.4.0"
yookassa = "^3.3.0"
black = "^24.8.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
